Kernel: Overhaul signal handling
- Fix race conditions with signals
- Support SIGSTP/SIGCNT
- Proper wait(2) return values

+++ kernel/syscall/kill.cpp
@@ -17,25 +17,23 @@ int Process::sys_kill(pid_t pid, int sig) {
 		auto* procs = TaskManager::process_list();
 		for(int i = 0; i < procs->size(); i++) {
 			auto c_proc = procs->at(i);
-			if((_user.uid == 0 || c_proc->_user.uid == _user.uid) && c_proc->_pgid == _pgid && c_proc->_pid != 1)
+			if(c_proc != this && (_user.uid == 0 || c_proc->_user.uid == _user.uid) && c_proc->_pgid == _pgid && c_proc->_pid > 1)
 				c_proc->kill(sig);
 		}
-		kill(sig);
 	} else if(pid == -1) {
 		//kill all processes for which we have permission to kill except init
 		auto* procs = TaskManager::process_list();
 		for(int i = 0; i < procs->size(); i++) {
 			auto c_proc = procs->at(i);
-			if((_user.uid == 0 || c_proc->_user.uid == _user.uid) && c_proc->_pid != 1)
+			if(c_proc != this && (_user.uid == 0 || c_proc->_user.uid == _user.uid) && c_proc->_pid > 1)
 				c_proc->kill(sig);
 		}
-		kill(sig);
 	} else if(pid < -1) {
 		//Kill all processes with _pgid == -pid
 		auto* procs = TaskManager::process_list();
 		for(int i = 0; i < procs->size(); i++) {
 			auto c_proc = procs->at(i);
-			if((_user.uid == 0 || c_proc->_user.uid == _user.uid) && c_proc->_pgid == -pid && c_proc->_pid != 1)
+			if((_user.uid == 0 || c_proc->_user.uid == _user.uid) && c_proc->_pgid == -pid && c_proc->_pid > 1)
 				c_proc->kill(sig);
 		}
 		kill(sig);
@@ -44,7 +42,7 @@ int Process::sys_kill(pid_t pid, int sig) {
 		auto proc = TaskManager::process_for_pid(pid);
 		if(proc.is_error())
 			return -ESRCH;
-		if((_user.uid != 0 && proc.value()->_user.uid != _user.uid) || proc.value()->_pid == 1)
+		if((_user.uid != 0 && proc.value()->_user.uid != _user.uid) || proc.value()->_pid <= 1)
 			return -EPERM;
 		proc.value()->kill(sig);
 	}
